pyproject.toml
[project]
name = "rfid-gate-log"
version = "0.1.0"
description = ""
authors = [
    {name = "Anders Birkenes",email = "datagutten@datagutten.net"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=6.0,<7.0)",
    "feig-py @ git+https://github.com/datagutten/rfid-gate-py.git",
    "requests (>=2.32.5,<3.0.0)",
    "sip2 @ git+https://github.com/tzeumer/SIP2-Client-for-Python.git",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "typenv (>=0.2.0,<0.3.0)",
    "flask (>=3.1.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
